requirements.txt
attrs==23.2.0
boltons==21.0.0
bracex==2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
defusedxml==0.7.1
exceptiongroup==1.2.0
face==22.0.0
glom==22.1.0
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
peewee==3.17.1
prompt-toolkit==3.0.43
psutil==5.9.8
Pygments==2.17.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semgrep==1.56.0
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
wcmatch==8.5.1
wcwidth==0.2.13
